Added Generation Action Functionality
A user can now define an action to be taken after a certain action is
parsed by the quest generator.

Example use case: a quest generates a store keeper who the player needs
to buy a certain item from. A generationAction could be defined that
places the generated item into the store keepers inventory.