Refactored parseArgument method
This has actually given a little bit more freedom in our grammar for
defining strategies. We do not HAVE to start with an array access. If
the user wants the argument to parse as the array itself, they can just
stop at the array identifier. Any id past that will be parsed as a
property access however, which could lead to strange behavior.